+++ pyproject.toml
@@ -23,6 +23,7 @@ dependencies = [
     "pandas",
     "matplotlib",
     "tabulate>=0.9.0",
+    "typing_extensions>=4.9"
 ]
 
 [project.urls]
